move resolvent methods into context (you saw that coming, right?)
